Final Commit?
Made the height fixed on the desktop version for the reviews, this stops the list from going different heights in the other rows. It also leaves a lot of white space though but I think its the better of the two. Because of the white space this created I split the class that was shared with the cart suggestions and the review list, this is so the cart suggestion list didn't have all the white space that was needed for the reviews. Updated some of the text so its not all just paragraphs on some pages.